requirements.txt

Flask==3.0.3
google-genai
pydantic
gunicorn